add fisheye_mask_generator with FoV & pole-polygon masking. Tested.
The masking executable takes in: a raw image, a camera intrinsics config file, a desired FoV, and output mask image path.

The masking has two steps: FoV masking and pole-polygon masking.

FoV masking takes in a desired FoV angle and masks all pixels that are out of this FoV. Since using fisheye, spherical normalization is used to normalize each pixel. This step typically generates a circular mask. Note that pixels on the corner may be false positive (retained), I think this is due to super large distortion. Now I just use eraser, but in the future it should be removed automatically (perhaps with flooding from image center).

The pole-polygon asks user to click a polygon shape around the laser pole to remove that region from the mask. It gives a FoV-masked original image captured by the camera. As the user clicks, polygon edges are drawn. Press 'y' to finish the clicking process.

Finally the raw image overlaid with red mask is visualized. The binary mask file is also saved as a png image to the user-specified path.